Refactor: extracted classes out of DialectInstanceParser
