+++ shared/src/main/scala/amf/plugins/document/vocabularies/parser/instances/DialectInstanceParser.scala
@@ -36,292 +36,6 @@ import org.yaml.model._
 
 import scala.collection.mutable
 
-trait NodeMappableHelper {
-
-  def allNodeMappingIds(mapping: NodeMappable): Set[String] = mapping match {
-    case nodeMapping: NodeMapping           => Set(nodeMapping.id)
-    case unionNodeMapping: UnionNodeMapping => unionNodeMapping.objectRange().map(_.value()).toSet
-  }
-
-}
-
-class DialectInstanceDeclarations(var dialectDomainElements: Map[String, DialectDomainElement] = Map(),
-                                  errorHandler: ErrorHandler,
-                                  futureDeclarations: FutureDeclarations)
-    extends VocabularyDeclarations(Map(), Map(), Map(), Map(), Map(), errorHandler, futureDeclarations)
-    with NodeMappableHelper {
-
-  /** Get or create specified library. */
-  override def getOrCreateLibrary(alias: String): DialectInstanceDeclarations = {
-    libraries.get(alias) match {
-      case Some(lib: DialectInstanceDeclarations) => lib
-      case _ =>
-        val result =
-          new DialectInstanceDeclarations(errorHandler = errorHandler, futureDeclarations = EmptyFutureDeclarations())
-        libraries = libraries + (alias -> result)
-        result
-    }
-  }
-
-  def registerDialectDomainElement(name: String,
-                                   dialectDomainElement: DialectDomainElement): DialectInstanceDeclarations = {
-    dialectDomainElements += (name -> dialectDomainElement)
-    if (!dialectDomainElement.isUnresolved) {
-      futureDeclarations.resolveRef(name, dialectDomainElement)
-    }
-    this
-  }
-
-  def findDialectDomainElement(key: String,
-                               nodeMapping: NodeMappable,
-                               scope: SearchScope.Scope): Option[DialectDomainElement] = {
-    val nodeMappingIds = allNodeMappingIds(nodeMapping)
-    findForType(key, _.asInstanceOf[DialectInstanceDeclarations].dialectDomainElements, scope) collect {
-      case dialectDomainElement: DialectDomainElement if nodeMappingIds.contains(dialectDomainElement.definedBy.id) =>
-        dialectDomainElement
-    }
-  }
-
-  def findAnyDialectDomainElement(key: String, scope: SearchScope.Scope): Option[DialectDomainElement] = {
-    findForType(key, _.asInstanceOf[DialectInstanceDeclarations].dialectDomainElements, scope) collect {
-      case dialectDomainElement: DialectDomainElement => dialectDomainElement
-    }
-  }
-
-  override def declarables(): Seq[DialectDomainElement] = dialectDomainElements.values.toSet.toSeq
-}
-
-class DialectInstanceContext(var dialect: Dialect,
-                             private val wrapped: ParserContext,
-                             private val ds: Option[DialectInstanceDeclarations] = None)
-    extends ParserContext(wrapped.rootContextDocument, wrapped.refs, wrapped.futureDeclarations, wrapped.eh)
-    with SyntaxErrorReporter {
-
-  var isPatch: Boolean                                           = false
-  var nestedDialects: Seq[Dialect]                               = Nil
-  val libraryDeclarationsNodeMappings: Map[String, NodeMappable] = parseDeclaredNodeMappings("library")
-  val rootDeclarationsNodeMappings: Map[String, NodeMappable]    = parseDeclaredNodeMappings("root")
-
-  def computeRootProps: Set[String] = {
-    val declarationProps: Set[String] = Option(dialect.documents()).flatMap(_.declarationsPath().option()) match {
-      case Some(declarationsPath) => Set(declarationsPath.split("/").head)
-      case _ =>
-        (
-          Option(dialect.documents()).flatMap(d => Option(d.root())).map(_.declaredNodes().map(_.name().value())).getOrElse(Seq()) ++
-            Option(dialect.documents()).flatMap(d => Option(d.library())).map(_.declaredNodes().map(_.name().value())).getOrElse(Seq())
-        ).toSet
-    }
-    declarationProps ++ Seq("uses", "external").toSet
-  }
-
-  var rootProps: Set[String] = computeRootProps
-
-  globalSpace = wrapped.globalSpace
-
-  def forPatch(): DialectInstanceContext = {
-    isPatch = true
-    this
-  }
-
-  def registerJsonPointerDeclaration(pointer: String, declared: DialectDomainElement): Unit =
-    globalSpace.update(pointer, declared)
-
-  def findJsonPointer(pointer: String): Option[DialectDomainElement] = globalSpace.get(pointer) match {
-    case Some(e: DialectDomainElement) => Some(e)
-    case _                             => None
-  }
-
-  val declarations: DialectInstanceDeclarations =
-    ds.getOrElse(new DialectInstanceDeclarations(errorHandler = eh, futureDeclarations = futureDeclarations))
-
-  def withCurrentDialect[T](tmpDialect: Dialect)(k: => T): T = {
-    val oldDialect = dialect
-    dialect = tmpDialect
-    rootProps = computeRootProps
-    val res = k
-    dialect = oldDialect
-    rootProps = computeRootProps
-    res
-  }
-
-  protected def parseDeclaredNodeMappings(documentType: String): Map[String, NodeMappable] = {
-    val declarations: Seq[(String, NodeMappable)] = Option(dialect.documents())
-      .flatMap { documents =>
-        // document mappings for root and libraries, everything that declares something
-        val documentMappings: Option[DocumentMapping] = if (documentType == "root") {
-          Option(documents.root())
-        } else {
-          Option(documents.library())
-        }
-        documentMappings.map { mapping =>
-          mapping.declaredNodes() map { declaration: PublicNodeMapping =>
-            findNodeMapping(declaration.mappedNode().value()) map { nodeMapping =>
-              (declaration.name().value(), nodeMapping)
-            }
-          } collect { case Some(res: (String, NodeMappable)) => res }
-        }
-      }
-      .getOrElse(Nil)
-
-    declarations.foldLeft(Map[String, NodeMappable]()) {
-      case (acc, (name, mapping)) =>
-        acc + (name -> mapping)
-    }
-  }
-
-  def findNodeMapping(mappingId: String): Option[NodeMappable] = {
-    dialect.declares.collectFirst {
-      case mapping: NodeMappable if mapping.id == mappingId => mapping
-    }
-  }
-
-  private def isInclude(node: YNode) = node.tagType == YType.Include
-
-  private def isIncludeMap(node: YNode): Boolean =
-    node.value.isInstanceOf[YMap] && node.as[YMap].key("$include").isDefined
-
-  def link(node: YNode): Either[String, YNode] = {
-    node match {
-      case _ if isInclude(node)    => Left(node.as[YScalar].text)
-      case _ if isIncludeMap(node) => Left(node.as[YMap].key("$include").get.value.as[String])
-      case _                       => Right(node)
-    }
-  }
-}
-
-case class ReferenceDeclarations(references: mutable.Map[String, Any] = mutable.Map())(
-    implicit ctx: DialectInstanceContext) {
-  def +=(alias: String, unit: BaseUnit): Unit = {
-    references += (alias -> unit)
-    // useful for annotations
-    unit match {
-      case vocabulary: Vocabulary => ctx.declarations.registerUsedVocabulary(alias, vocabulary)
-      case _                      =>
-    }
-    // register declared units properly
-    unit match {
-      case m: DeclaresModel =>
-        val library = ctx.declarations.getOrCreateLibrary(alias)
-        m.declares.foreach {
-          case dialectElement: DialectDomainElement =>
-            val localName = dialectElement.localRefName
-            library.registerDialectDomainElement(localName, dialectElement)
-            ctx.futureDeclarations.resolveRef(s"$alias.$localName", dialectElement)
-          case decl => library += decl
-        }
-      case f: DialectInstanceFragment =>
-        ctx.declarations.fragments += (alias -> FragmentRef(f.encodes, f.location()))
-    }
-  }
-
-  def +=(external: External): Unit = {
-    references += (external.alias.value()                 -> external)
-    ctx.declarations.externals += (external.alias.value() -> external)
-  }
-
-  def baseUnitReferences(): Seq[BaseUnit] =
-    references.values.toSet.filter(_.isInstanceOf[BaseUnit]).toSeq.asInstanceOf[Seq[BaseUnit]]
-}
-
-case class DialectInstanceReferencesParser(dialectInstance: BaseUnit, map: YMap, references: Seq[ParsedReference])(
-    implicit ctx: DialectInstanceContext) {
-
-  def parse(location: String): ReferenceDeclarations = {
-    val result = ReferenceDeclarations()
-    parseLibraries(dialectInstance, result, location)
-    parseExternals(result, location)
-    references.foreach {
-      case ParsedReference(f: DialectInstanceFragment, origin: Reference, None) => result += (origin.url, f)
-      case _                                                                    =>
-    }
-    if (ctx.isPatch) {
-      references.foreach {
-        case ParsedReference(f: DialectInstance, origin: Reference, None) => result += (origin.url, f)
-        case _                                                            =>
-      }
-    }
-
-    result
-  }
-
-  private def target(url: String): Option[BaseUnit] =
-    references.find(r => r.origin.url.equals(url)).map(_.unit)
-
-  private def parseLibraries(dialectInstance: BaseUnit, result: ReferenceDeclarations, id: String): Unit = {
-    val parsedLibraries: mutable.Set[String] = mutable.Set()
-    map.key(
-      "uses",
-      entry => {
-        val annotation: Annotation =
-          AliasesLocation(Annotations(entry.key).find(classOf[LexicalInformation]).map(_.range.start.line).getOrElse(0))
-        dialectInstance.annotations += annotation
-        entry.value
-          .as[YMap]
-          .entries
-          .foreach(e => {
-            val alias: String = e.key.as[YScalar].text
-            val url: String   = library(e)
-            target(url).foreach {
-              case module: DeclaresModel =>
-                parsedLibraries += url
-                collectAlias(dialectInstance, alias -> (module.id, url))
-                result += (alias, module)
-              case other =>
-                ctx.eh.violation(DialectError, id, s"Expected vocabulary module but found: '$other'", e) // todo Uses should only reference modules...
-            }
-          })
-      }
-    )
-    // Parsing $refs to libraries
-    references.foreach {
-      case ParsedReference(lib: DialectInstanceLibrary, _, _)
-          if !parsedLibraries.contains(lib.location().getOrElse(lib.id)) =>
-        result += (lib.id, lib)
-      case _ => // ignore
-    }
-  }
-
-  private def library(e: YMapEntry): String = e.value.tagType match {
-    case YType.Include => e.value.as[YScalar].text
-    case YType.Map if e.value.as[YMap].key("$include").isDefined =>
-      e.value.as[YMap].key("$include").get.value.as[String]
-    case _ => e.value
-  }
-
-  private def parseExternalEntry(result: ReferenceDeclarations, entry: YMapEntry): Unit = {
-    entry.value
-      .as[YMap]
-      .entries
-      .foreach(e => {
-        val alias: String = e.key.as[YScalar].text
-        val base: String  = e.value
-        val external      = External()
-        result += external.withAlias(alias).withBase(base)
-      })
-  }
-  private def parseExternals(result: ReferenceDeclarations, id: String): Unit = {
-    map.key(
-      "external",
-      entry => parseExternalEntry(result, entry)
-    )
-
-    map.key(
-      "$external",
-      entry => parseExternalEntry(result, entry)
-    )
-  }
-
-  private def collectAlias(aliasCollectorUnit: BaseUnit,
-                           alias: (Aliases.Alias, (Aliases.FullUrl, Aliases.RelativeUrl))): BaseUnit = {
-    aliasCollectorUnit.annotations.find(classOf[Aliases]) match {
-      case Some(aliases) =>
-        aliasCollectorUnit.annotations.reject(_.isInstanceOf[Aliases])
-        aliasCollectorUnit.add(aliases.copy(aliases = aliases.aliases + alias))
-      case None => aliasCollectorUnit.add(Aliases(Set(alias)))
-    }
-  }
-}
-
 class DialectInstanceParser(root: Root)(implicit override val ctx: DialectInstanceContext)
     extends BaseSpecParser
     with PlatformSecrets
